[mk/install] Warn if files from a PLIST are already present before install
Before installing a package, warn if the installation is going to overwrite
existing files. This does not normally happen (because conflicting packages
must be deleted before installation), but can still happen if:
 - original checksum of files have changed: robotpkg_delete leaves them behind.
 - manual installation of a package under LOCALBASE has created new files that
   are unknown to robotpkg.
 - two packages have a conflicting PLIST whitout being marked as conflicting
   packages.

In all those situations, it is interesting to detect and report the problem.

Doing 'make install confirm' (or update/replace confirm) will bypass this
check.