[pkgtools/pkg_install] Fix "large file" compilation options
pkg_install is linked with libarchive and must thus understand off_t the same
way as libarchive. This is done in pkgsrc by enforcing _FILE_OFFSET_BITS=64 as
well as various other macros, something which was not propagated in robotpkg
Makefile by mistake. However, it's imho more robust to use the AC_SYS_LARGEFILE
autoconf macro to fix this (this is at least consistent with what libarchive
does).

This problem affected robotpkg_add and robotpkg_create tools on Linux (which
are apparently not used much since they've probably been broken for years ...).

Bumped PKGREVISION.