[mk] Fix deinstall target when DEINSTALLDEPENDS=all
The deinstall target was still using variables from old-style dependencies to
deinstall build depends of a package (and thus was not deinstalling any build
depends when asked to do so). Update this to use new style variables.

While here, update printed messages: no need to bother the user with cryptic
shell commands, there is PKG_DEBUG_LEVEL for that. Instead, print the list of
packages that will actually be removed.