[mk] Implement 'bulk' target and enhance sets to provide a bulk-aware target
The 'bulk' target is a public target to build a binary package, and remove
it (together with any depending packages) immediately after that, leaving
only the binary package back. Any leftover files are reported.

Before building a package, bulk will install any robotpkg dependency from
available binaray packages. It can recursively invoke bulk for depending
packages, but this target is most interesting when used on a package set so
that a bunch of packages are built/checked at once.