[mk/update] Fix an error case when resintalling a dependent pkg during update
During the reinstallation of each removed dependencies, `make update` now
checks that a given package has not already been reinstalled*. If so, and if
the package was effectively recompiled with the correct dependency**, it simply
skips it.

* A package can have been already resinstalled either because:
 - dependencies are not tsorted, so another pkg in the update list might have
   required it already.
 - the user may have manually resintalled it because of a previous failure
   he is now resuming the update.
 - a package A may have been deleted with -f, leaving its parents alone and not
   deinstalled by `make deinstall` in A

** If a package is already present, one must check that is was actually
compiled with the exact version of the package being updated, because of the
third point above.