[mk/internal] Add new 'dewey' awk script dealing with pkg patterns comparisons
dewey.awk is the fusion of two functionalities used at different places in
robotpkg Makefiles.
 1. match a list of package patterns (e.g. x>=a.b) against a specific version
    and return a yes/no answer. (this is redundant with `pkg_admin pmatch`, but
    because of 2. below this functionality is present for free).
 2. take a list of patterns and reduce it to a single interval of the form
    >a<b. (!= constraints are handled specifically). (this is a much enhanced
    version of the now obsolete mk/depends/reduce-depends.awk).

dewey.awk understands the same set of patterns as pkg_admin and pkg_info since
the code converting a version string into an array of integers is the exact
copy of the C version in pkg_install/lib/dewey.c (except for the "nb" pkg
revision number which is "r" in robotpkg).

The main motivation for writing this is to get rid of horrid bits in the
Makefiles that were trying to implement this in gmake syntax. It should also
relax the dependency on pkg_admin tools during early stages of Makefile
parsing and maybe lessen the importance of the bootstrap script.