[mk] Rewrote the working directory cookie system to make it more dynamic
In order to make the working directory state less static and make dandling
working directories less problematic, the cookie system for each step has a
proper dependency registered on the previous state and can clean itself when
needed, to perform its task again. In addition, invoking a main target
(depends, configure, build, install) will unconditionnaly perform the task
again (extract and patch won't do that, of course). 'install' will also delete
an existing instance of itself before installing (but not recursively, so
nothing else will be removed - there's update/replace for that).

The base cookies (bootstrap-depends and depends) also have a dependency on the
SYSTEM_FILES of each dependency of a package, so that if something listed there
changes in the system, a package with a dandling working directory will be able
to reinvoke the depends/configure/build step transparently. This is of course
not so robust because a dependency is registered only on the few files from
SYSTEM_SEARCH, not the whole (system-) packages, and this also depends on the
underlying build system of the package being able to correctly track
dependencies etc. But all in all it seems better than doing nothing and/or fail
mysteriously. The best is still to 'make clean' and avoid keeping old work/ dir
around.

This change also impact the 'barrier' target, that is now fully implemented via
the gmake "Makefile remaking" feature and works globally better, especially
when specifying multiple targets on the command line. The cost is two extras
make 'restarts' vs. one recursive make before (there is one make restart for
the bootstrap depends, one for the depends). The benefit is that all dependency
information (PREFIX.<pkg>, SYSTEM_FILES.<pkg> ...) is directly available in a
package Makefile or depend.mk (is used to be available only in Makefiles
rules), and e.g. bootstrap depends can be used to tweak a regular
depend.mk. For instance, it will be possible to look for a python package for
just the specific version of python that would have been found during bootstrap
depends.