[mk/package] Merge package and tarup targets
The two targets 'package' and 'tarup' invoke the same rules, which the only
difference that 'package' fails if the latest version of a package is not
installed and 'tarup' will create a package from any installed version.

In addition, both targets will do a recursive make install if the package is
not installed at all.