[mk/update] Fix 'update' target when resuming an update
When resuming an update, the package being updated was systematically
reinstalled. This used to work because 'make install' used to be a no-op for
already installed package, but now that 'make install' is never a no-op, this
breaks.

Simply testing for the package existence before recursing in install (as for
the dependent packages) fixes the problem.