[mk/sysdep] Compute TCLLIBPATH for packages requiring tcl packages.
A suitable value for TCLLIBPATH is computed for all packages that mention a
pkgIndex.tcl in their SYSTEM_SEARCH variable. A specific path can also be
forced by defining a TCLPATH.<pkg> variable (absolute path or relative to
PREFIX.<pkg>).

The default search path of tclsh (as reported by the $auto_path variable) is
never added to TCLLIBPATH.