[mk/sets] Add a predefined 'depends' set
The 'depends' sets merely maps to the current package (therefore only
available when the cwd is inside a package). It is defined as not 'strict', so
that it can be used to work on a package and all of its dependencies.

Removed the specific 'clean-depends' implementation that is now useless.