[mk/fetch] Add missing include for DISTINFO definition
When invoking higher-level targets than fetch, the fetch target was not
verifying the distfile against the recorded checksum because of a missing
include of checksum-vars.mk. The checksum was still verified, but only right
before extraction. At this point, if the archive happens to be corrupted, the
fallback to the backup download site is not available anymore.