[pkgtool/pkg_install] Make ROBOTPKG_BASE and LOCALBASE truly independent
ROBOTPKG_BASE and LOCALBASE can now safely point to different locations, even
temporarily on the command line. For instance one can quickly install packages
to a different prefix on the command line without bootstraping to that prefix,
with e.g  `make update LOCALBASE=/tmp`, while still using robotpkg_* tools in
ROBOTPKG_BASE (or ${PATH}).

If ROBOTPKG_BASE is not equal to LOCALBASE, then pkgtools/pkg_install is
considered as a 'system' package. This can still be overriden if needed with
PREFER.pkg_install as usual.