[mk] Workaround variables in MAKEFLAGS sometimes not passed to recursive make
For some obscure reason, variables passed on the command line are lost after
two recursions of gmake-3.82 (e.g. when installing a dependency of a package),
when gmake restarts itself after bootstrap dependency resolution.
See http://savannah.gnu.org/bugs/index.php?33129 for a simple Makefile showing
the problem.

Since all recursive invocation of ${MAKE} in robotpkg actually use the
${RECURSIVE_MAKE} variable, this can be worked around with little effort...

While here, remove a useless export of MAKECONF: if it was specified on the
command line, it will be always defined (and ignored in the environment).