[mk/depends] Let the 'depends' target depend on all included Makefiles
If a package Makefile (or one of its dependency) is modified and a package
working directory is not clean, this is worth re-checking the dependencies of
this package. To do this, let the 'depends' and 'bootstrap-depends' cookies
depend on all included Makefiles but those in WRKDIR (because those in WRKDIR
are typically cookies from later targets).

The files included by robotpkg.conf are now automatically included in this
dependency list, so do not bother storing them in a specific variable.