[mk/pkg] Fix up-to-date checks for 'update' and 'replace' targets
Rewrite the tests computing the up-to-date status of a package, preventing a
make update or replace without confirmation to deinstall a package if this is
not strictly required.

Instead of using fragile $(shell) constructs evaluated during Makefile parsing,
use a proper target and rely on gmake included files remaking to trigger this
rule when appropriate. The tests can now be more complex and thus more
accurate. In particular, a package is considered outdated if its dependencies
have changed or its PKG_OPTIONS were modified.

Notice that this does not interfere with the real dependencies checking and
the tests performed here need not to be 100% accurate. This target is merely a
helper target trying to prevent mistyped commands to start upgrading a lot of
packages when this was not wanted.