[mk/update] Fix update recovery after a failure for a dependency reinstallation
The update target skips already reinstalled dependencies when it is restarted
after a failure. However, it was checking for a reinstalled package according
to its old name before the initial update. This package may have a new version
and one should avoid testing for a specific version in this case.

This commit adds one more place where it is assumed that a package version
contains no dash.