[mk/platform] Introduce SYSLIBDIR to account for new system lib dirs of ubuntu
Ubuntu-11.04 has randomly moved some libraries in lib/{i386,x86_64}-linux-gnu.
The SYSLIBSUFFIX variable that was accounting for lib vs. lib64 needs to be
replaced by the new SYSLIBDIR that contains a list of directories where system
libraries can be found.

The algorithm finding a system library is as follow: if SYSLIBDIR is
defined, every file specification in a SYSTEM_SEARCH.pkg that starts with lib/
is replaced by a pattern matching _existing_ directories of SYSLIBDIR in the
current search prefix (thus lib/ is not searched except if it is explicitely
mentionned in SYSLIBDIR). If no directory from SYSLIBDIR exist in the current
prefix, lib/ is left unchanged.

Currently, SYSLIBDIR is undefined on anything but Linux, and default to
'undefined' for 32bits linux and 'lib64' for 64bits machines. On ubuntu,
SYSLIBDIR is set to "lib/${MACHINE_ARCH}-linux-gnu lib", or just "lib" if no
"/usr/lib/${MACHINE_ARCH}-linux-gnu" exists.