[image/openni-nite] Fix MACHINE_ARCH undefined
Include robotpkg.prefs.mk to have MACHINE_ARCH defined before defining the
roper MASTER_SITE. Also split MASTER_SITE into elements that fit on 80 columns.
And while here, removed a few 'space-before-tab' whitespace.

IMPORTANT NOTE: the archive currently referenced by the package (1.3.0.17) is
NOT AVAILABLE ANYMORE for download. Since the archive may not be
redristributed, it cannot be mirrored on MASTER_SITE_BACKUP. So this package
should probably be updated to the newer 1.3.1.3, but I don't know how much
things this would break, so I leave it as is for now.