[mk/depends] Let the 'depends' target depend on 'cbbh'
The depends target normally depends on bootstrap-depends that in turn depends
on cbbh. However, when bootstrap-depends was already checked in a previous run,
cbbh might not be (re)checked. This is a problem if e.g. the package options
have been changed in the meantime, leading to e.g. a 'make clean' failing with
an 'please do a make clean' error.