[devel/swig] Import swig-2.0.4
SWIG (Simplified Wrapper and Interface Generator) is a compiler that integrates
C and C++ with languages including Perl, Python, Tcl, Guile, Mzscheme, Java,
Ruby, PHP, and Ocaml.

This package is made available in robotpkg because version 2.0.4 is not widely
available yet and 2.0.4 has important fixes for python as well as support for
python-3.2.

The mk/sysdep/swig.mk file was moved to devel/swig/depend.mk and packages using
it have been fixed accordingly.