[mk] Put an end to the recursive definition of PKGNAME and PKGREVISION
PKGNAME and PKGREVISION are special variables in that they may be defined by a
package, but need to be suffixed by PKGREVISION (if present) automatically.

Doing this by using recursive ':=' assignment raises an issue if the original
definition of PKGNAME contains variable references (like the upcoming
PYPKGPREFIX). The recursive assignment looses the initial variable references
and expands them (usually too early, before they are actually defined).

So assign PKGNAME and PKGVERSION with a combination of $(eval .. $(value ..))
that prevent that early expansion to occur.