[mk] Put an end to the 'make clean' => 'please do a make clean' issue
Under certain circumstances, `make clean` would fail and trigger an error
asking for a 'make clean' (see 489b1072). This is due to 'make' trying to
rebuild included dependency information before the actual 'clean' target can
run.

The fix in 489b1072 was OK but incomplete, because not only depend/depend.mk
requires bootstrap dependency information, but virtually all targets
implicitely depend on this. So the proper fix is to include files that need to
be rebuilt (and depend on cbbh) only when the package can be built here, and do
nothing otherwise. Then the 'cbbh' target will run only when needed (e.g. not
for a `make clean`).