[mk] Introduce DEPEND_VARS.pkg in depend.mk
depend.mk may now set DEPEND_VARS.pkg to a list of variable names.
DEPEND_VARS.pkg should only be appended to (always use +=).

When set, DEPEND_VARS makes a _parent_ package (including the depend.mk in
question) export the value of the variables to the recursive make process of
the child packages (thus only if not installed yet and not a system pkg).

DEPEND_VARS cannot be used to pass options or anything affecting the build
process itself, since it will not be used if the package is already
installed. DEPEND_VARS should probably be used only for passing an
'ALTERNATIVE.pkg' definition to require a dependency with the same alternative
selection as its parent (and again this will work only if the child package has
a variable part in its name depending on this alternative).

Technically, DEPEND_VARS is implemented by setting environment variables of the
form _override.<pkgpath> = <var list> and _override.<var> = <value> so that any
recursive make in the <pkg> directory will be able to reconstruct the variable
definitions. (this is complex but required by targets like 'show-depends' that
invoke MAKE in several unrelated directories in any order). The
env. var. filtering rules have been updated to not filter those variables.