[mk] check-depends target requires bootstrap-depends to run correctly
Strictly speaking, check-depends needs to run after bootstrap-depends so that
dependencies that require a bootstrap phase are correctly handled. Currently,
no package really needs this, but that's probably just a matter of time...

This breaks a little bit the assumption that show-% targets will not need a
WRKDIR, so for now 'show-depends' invokes 'check-depends' for child packages
with a temp WRKDIR pointing inside the WRKDIR of the package where show-depends
is invoked.

While here, robustify the output of 'check-depends' so that regular messages
and actual dependency information are more easly sorted out, especially since
check-depends now outputs the messages from bootstrap-depends.