[mk] Fix bootstrap-depends or depends invoked twice on failures
Including dependency information files .{,sys}{,bs}depends with '-include' has
the nasty side effect of triggering a 'make restart' because even if the files
are initially non-existent, gmake has recorded a dependency on them and detects
they have been rebuilt by {,bootsrap-}depends. This is especially bad if the
{,bootsrap-}depends fails because then {,bootsrap-}depends gets called twice
(at each 'make restart').

Simply using 'include $(realpath files...)' has the same effect as -include
except that this does not trigger gmake restart (gmake sees an empty include if
the files are not present). As an extra bonus, this makes it useless to
define empty rules for those files.