[mk] Fix 'depends' invoked twice when 'bootstrap-depends' is outdated
The depends target is deferred until bootstrap-depends has run. If
bootstrap depends has run already, one must still check before running depends
that bootstrap-depends is not outdated and won't be rebuilt upon next restart.

The fix is a bit tricky and relies on the fact that gmake does not display
errors while remaking included Makefiles (and treat them as non-fatal for the
actual cmdline targets).