[pkgtools/pkg_install] Improve the situation when PREFER.pkg_install == system
It is not necessary to use the newly built pkg_install tools in the Makefile
when PREFER.pkg_install == system, so restrict the PKG_*_CMD overloading to the
PREFER.pkg_install == robotpkg case.

On the other hand, deinstalling a pkg_install when PREFER.pkg_install == system
is allowed, so hook to pkg_deinstall target to force removal. The PKG_PRESERVE
flag is still enforced, though, so that the binary package has the attribute
set.