[mk] Implement the "~<tags>" matching in awk dewey script
mk/internal/dewey.awk script needs to know about the ~<tags> to correctly
reduce a set of package patterns. While here, also add it to the 'pmatch'
subcommand, although it is not currently used (pkg_admin pmatch is used).

Reduction for options is not robust as soon as it comes to glob patterns. It
should work for basic cases, though. Implementing a real option reduction is
too complex at the moment.