[mk] Add current option list to the package name
PKGNAME is redefined to include the options used to build the package. This is
done by using the fresh ~<tags> feature from pkg_install. There should not be
any visible user changes, apart from the fact that robotpkg_* commands will
list slightly different package names.

The idea behind this change is to let multiple binary packages archives coexist
in ${PACKAGES} and also properly encode required options in the dependencies of
those archives.

Notice that this breaks the assumption that ${pkg%-*} is the package base name,
as option names may contain dashes. One must use ${pkg%~*} and then ${pkg%-*}
to retrieve the base name from the shell.