[pkgtools/pkg_install] Do not reinvent pkg_match() for matching a pkg basename
Packages may be given by basename (i.e. without version) on the pkg_info et al.
command line. In this case, the operation is performed on any installed package
matching the given basename. Instead of reinventing the matching algorithm, use
pkg_match() on the result of the new pkgaddwildcard() function.