[mk] Automatically derive PKG_ALTERNATIVE.<alt> settings from a reqd pkgname
Add the command-line variable PKGREQD that let a package automatically select
alternative settings (if any) from the PKGTAG.<alt> strings. PKGREQD actually
contains the DEPEND_ABI required by parent packages, so this will be used later
to compute option settings as well. This make DEPEND_VARS.<pkg> obsolete.

Note:
It is probably less robust to match a random string in the package name
than explicitely exporting settings via DEPEND_VARS, but DEPEND_VARS was not
generic enough to handle on the one hand PKG_OPTIONS and on the other hand a
simple dependency on a specific alternative (e.g. depend only on viam-tcl-genom
and not any other alternative when 'tcl' is an alternative of 'viam-genom').

Using PKGREQD will work well as long as a package name does not have a
substring matching any PKGTAG defined by alternatives (e.g. a package using
python cannot be called py26-py27-pkg...). This is also the way pkgsrc handles
this issue, so this should work on the long term... :)