[mk] Enhance sets definition with wildcards.
Sets can now be defined with wildcards in the path and/or package name. The
general format is as follow:
        category/subdir:pkgname

category/subdir can contain usual shell globbing characters, that are expanded
according to the SUBDIRS variable in toplevel and category Makefiles.

pkgname can also contain glob characters, that are expanded according to the
result of the 'print-pkgnames' target. Options are expanded only if a ~
character followed by some wildcards is present.

For instance, the pattern */*:*~* will expand to all possible packages,
alternatives and options combinations (in about ~20mins on NFS, generating
8000+ package names ;).

In case of different patterns applying to the same package, only the first
pattern that generates a match is taken into account. This allows to define
strict rules that apply only to some packages. E.g. 'devel/*:*~!debug
*/*:*~debug' will expand to all packages without the debug option in devel, and
with the debug option in any other subdir.