[mk] Teach dewey how to 'reduce' and 'getopts' with {,} alternatives
For 'reduce', do all alternatives recursively and merge the result as a single
alternative list. A naive optimization to find the largest common prefix and
suffix is then done to output a better looking result in simple cases.

For getopts, each alternative is analyzed in turn and the smallest non-null
option set generated by any of them is returned.