[mk] prefixsearch.sh can match simple options in a system package
Options in a system package that result in a different file or different file
content can be detected by passing a special file spec in SYSTEM_SEARCH.pkg
The file:sed:prog:opt format will look for 'file' and pass it to sed -n with
the 'sed' program, or invoke 'prog' and filter the output with the 'sed'
program. If this results in any output, the package will be considered to have
option 'opt'. E.g :
    include/foo.h:/#define OPT/p::bar
will report that it has found pkg~bar if foo.h contains the string #define OPT.

Since options are ... optional, a missing file with an option qualifier will
not be reported as an error unless a dependency requires this particular
option. A missing file will still be included in SYSTEM_FILES as the special
"/notfound" path (to keep the same number of elements in SYSTEM_SEARCH and
SYSTEM_FILES).

The /notfound path in SYSTEM_FILES.pkg must be filtered out being added as a
dependency of the ${_COOKIE.depends} file.