[mk] Adapt 'update' target to handle automatic option settings.
Updating a package now require an analysis of installed packages to propagate
automatic option settings/requirement that may be set be packages required by
the updated one. This is done by passing the 'required_by' list through the
tsort-set script, that will compute the exact package names to be reinstalled.