[pkgtools/pkg_install] Teach pkg_install how to compute OPSYS and OS_VERSION
Although OPSYS and OS_VERSION are statically compiled in pkg_install, pkg_add
needs to dynamically check for the OS_VERSION (that may be different from the
compiled in version). So embed a C version of mk/platform/opsys.sh in
normalise_platform().

Bump version to 20110805.2, as well as DEPEND_ABI.