[mk] Workaround an issue with NetBSD's /bin/sh + NFS work directories
NetBSD-5.1 /bin/sh has a problem with closing file descriptors used in a while
loop. Reorder things and use two different shells in 'update-clean' so that if
a work dir is on an NFS share, the dandling file descriptors do not cause
failures in "rm -rf".

How to repeat:
$ ls /proc/self/fd
0 1 2 3 4
$ while false; do :; done 9</dev/null
$ ls /proc/self/fd
0  1  10 2  3  4  9

See also:
http://mail-index.netbsd.org/tech-userlevel/2011/09/26/msg005466.html