[mk] Improve prefixsearch.sh error reporting
For SYSTEM_SEARCH patterns checking version, remember the exit status of the
command used to retrieve a version number.

If the exit status is 0 and the command did not output anything, assume that
there is no version instead of raising an error (other patterns in
SYSTEM_SEARCH will either find a version or there will be a warning printed at
the end).

If the exit status is not 0, use the output as an error message (in sysdep.log)
instead of the useless 'unknown version' message. Fall back to "error" if there
was no output.