[mk] Cache "not available" information for bulk target
An entry is added in ${PACKAGES}/NotAvail for each package that is not
available. The content of the file is the reason why the package is not
available, plus the chain of dependency, if any. Later, the bulk up-to-date
checks will not try this package again if the dependencies haven't changed (as
for "available" packages).