[mk/sysdep] When checking for flex, also check libfl
Fedora flex package is missing libfl (it is in flex-static). So make sure to
check both bin/flex and lib/libfl and update the system package names for the
usual OS.

While here, remove unused option 'flex' for selecting between lex and flex. If
this is ever actually needed, this should be done via a proper 'alternative'.