[cross/i386-mingw32] Remove package and make it a pure sysdep
Since i386-mingw32 is required for only one package (net/evart-stream), it's
not really worth to have it in robotpkg tree. Especially since all supported
systems now provide it with system packages.

Add mk/sysdep/mingw32-c for the mingw C compiler. C++ compiler is not needed at
this time, but could be added easily.

The selection of the mingw C compiler is still triggered by the hackish
COMPILER_TARGET = i386-mingw32 variable. This should be improved somehow, but
later ...