[mk/sysdep] Restrict available version of py-setuptools depending on OPSYS
Fedora 14 has py-setuptools for python26 and python31
Fedora 15 has py-setuptools for python27 and python32
NetBSD has py-setuptools for python26 and python27 (currently broken for 31)
Ubuntu has all the versions

This restriction is a bit abusive, since anyone is free to install
python-setuptools manually for the desired python version. But I think it's
better not to encourage this.