[mk] Pass -DPKGMANDIR and -DPKGINFODIR to cmake configure scripts
As with gnu-configure, manual pages and info pages should be installed into
robotpkg hier (i.e. prefix/info prefix/man by default).

Pass -DPKGMANDIR and -DPKGINFODIR to cmake configure scripts so that they can
use those variables if needed. This will trigger a "CMake Warning: The variable
was not used" most of the time, but I guess this is harmless?