[mk] Fix tsort-set.awk in a corner case situation
tsort-set.awk computes for each package of a set the available package names
and the dependencies of the package. This is done by invoking `make
print-pkgnames` and `make print-depends-pkgpaths`. In order to optimize this,
the two targets are invoked via a single make command.

However, the print-pkgnames target has the nasty side effect of changing both
PKG_OPTIONS and PKG_ALTERNATIVE.x for all x alternatives of the package, so
that it can mess up the actual dependencies of a package (e.g. instead of a
dependency on py26-xxx, it can introduce a wrong dependency on py32-xx because
of the modified alternatives).

A simple workaround is to invoke the print-pkgnames target last on the command
line. I could not find another solution (yet?) to really avoid the side
effects and keep the optimization that uses a single make command.