[image/artoolkitplus] Fix DOS line endings in local patches
There is not reason to keep DOS line endings in the local patches, since the
archive is extracted with unzip -a, which normally gets rid of this.

However, unzip -a seems to behave differently on different platforms (and unzip
in general, although it indentifies itself as unzip-6.0). So force text file
type for all files with -aa by locally overriding unzip options.