[mk] Rebuild a package in bulk if any of its dependencies has a newer version
Bulk used to check if a package was up-to-date by comparing the build date of
its dependencies to the build date of the current package. This is correct,
however one want bulk to rebuild the package if anything has changed (so even
if an old dependency still satifies the package). Instead of checking the
date of the exact dependency the package was built with, check the date of the
best dependency that satisfies the package.