[meta-pkgs/ros-base] Improve manual install rule
The list of files to install for ros packages is taken from the package PLIST.
This requires the local do-install target to be able to understand PLIST
variables. The best for this is to use the same rules as for plist generation.

Also add compilation of python files to the install target.