[middleware/ros-actionlib] Import ros-actionlib-1.6.1
The actionlib package provides a standardized interface interfacing with
preemptible tasks on top of the standard ROS middleware tools. Examples of this
include moving the base to a target location, performing a laser scan and
returning the resulting point cloud, detecting the handle of a door, etc.

Note that the package currently fetches the "common" stack, which contains a
few more packages than actionlib, but it only build actionlib. This is because
the "common" stack has been deprecated starting from 1.7 and replaced by a
stack with the single actionlib package.