[mk] Delay subst.mk inclusion as much as possible
subst.mk (In-place file substitutions) must be included as late as possible in
robotpkg.mk, because it performs immediate expansion on variables set in other
mk files (this is also true for su-target.mk). So include those files in the
footer of robotpkg.mk

This fixes an issue when invoking print-PLIST on a clean package that defines
substitutions, e.g. for a package using gnu-configure without C++
requirements, where the substitution in question disable fatal errors in
the configure script for CXXCPP. (this is by the way not the recommended
procedure for invoking print-PLIST: print-PLIST should be invoked after a make
install for best results).

Problem found and reported by Matthieu for robots/rmp400-genom.