[math/fsqp] Use standard interactive fetch message instead of custom one
The standard message is good enough to indicate that the distfile must be
requested from aemdesign. Furthermore, the custom message contained misleading
information as it advised to contact ${MAINTAINER}, who is not responsible
for distributing the distfile.

Note: www.aemdesign.com seems to have been down for a long time now. Not sure
about the future of this package...