[mk] Update system dependency resolution for system libraries not in /usr/lib
The simple strategy consisting in detecting if a sysdep file specification is a
system library by checking if the path starts with "lib/" breaks when the file
specification contains alternatives: {,x}lib/ will for instance fool the
script. So rewrite this part of the code to run after the brace expansion and
shell wildcard expansion, taking care of not expanding the patterns twice and
not breaking the optional /usr prefix removal for the special /bin and /lib
directories.

This should fix sysdep/qt4-libs.mk on ubuntu-11.10, problem reported by
adegroot@