[mk] Change the version suffix used by checkouts (use PKGREVISION)
The intent is to make checkout version numbers (i.e. the result of a `make
checkout`) more robust and fix various glitches that occur when a package
defines PKGVERSION itself instead of relying on the automatic computation of
this variable.

Checkouts now use the PKGREVISION variable instead of appending a custom tag to
the PKGVERSION variable. A typical 'checkout' version will now look like
pkg-1.0r020120119141105, where the big number is of the form
0YYYYMMDDHHMMSS. The leading 0 is here to give a visual indication that this is
a checkout, since regular PKGREVISION will normally not start with a 0. If the
package already has a PKGREVISION, the checkout tag is simply appended to the
existing PKGREVISION (although this is not critical to combine both, since
regular PKGREVISION and checkouts are somewhat redundant from the package
manager p.o.v.).

The resulting binary packages also have the build variable CHECKOUT_PKGVERSION
defined (see robotpkg_info -B).