[math/jrl-mathtools] Fix configure with cmake-2.8.6
cmake-2.8.6 has a buggy FindLAPACK: typo in CMAKE_FIND_LIBRARY_SUFFIXES
spelling (#12624, see http://public.kitware.com/Bug/view.php?id=12624).
Add a pre-configure rule to copy a working FindLAPACK.cmake and define
CMAKE_MODULE_PATH to point to this script.

Note that this fix does not really belong to this package, since this is a
cmake issue, but no other robotpkg package uses cmake's FindLAPACK and there is
not better place atm to handle this kind of issue.

This fix should be removed if/whenever Fedora16 people update to cmake-2.8.7 or
fix their cmake-2.8.6.