[mk/sysdep] Even more search directories for system python package
Every day sees a new flavour of installation path for python system packages
on ubuntu. Today, add '/usr/lib/pyshared/pythonx.y' to the search list, for
py-matplotlib on ubuntu.

The main problem being that during dependency resolution, PYTHON interperter is
not set yet, so it cannot be used to retrieve sys.path ...