[mk] Install binary dependency packages during bulk build with robotpkg_add -u.
This is needed for the following scenario:
If the package being built depends on packages "p" and "q". If package "p"
depends on "a" and a package "q" depends on "a~opt" (i.e. same package but
with some required options), the wrong "a" may be initially installed depending
on the order in which things are done. Since the installation order cannot be
(easily) controlled, using robotpkg_add -u will smoothly upgrade "a" to "a~opt"
if/when needed. If "p" cannot be installed with "a~opt" (e.g. "p" requires
"a~!opt"), robotpkg_add -u will still report the conflict and bulk will fail
properly.