[mk] Fix up-to-date check in bulk for packages changed from n/a to available
When bulk-building a package, one must not only clean any "not available" cache
file for the package, but also for all its versions. This is required because a
package can be bumped and updated from "not available" to "available" at the
same time. Failing to clean cache files for previous versions will let any
package depending on the upgraded one think they still can't be built.