[pkgtool/pkg-config] Update run-time path addition strategy
Patching pkg-config files for adding the run-time path directly in the source
tree raises some issues (e.g., finding the proper source files to patch, or
locally modified files when working with 'checkouts;).

Instead, patch the files directly in PREFIX, during the installation phase. The
default pattern for finding files to patch is lib/pkgconfig/*.pc. Note this
requires a valid PLIST, so patching won't work while developing a new package
when the PLIST is not yet created.

While here, also improve the patching program so that it does not duplicate any
existing rpath flags.