[architecture/genom] Define a common, generic PLIST for genom modules
Instead of duplicating basically the same PLIST for all genom modules, define
the generic part in a common file (PLIST.module). This saves space, typos and
prevent mistakes. While here, also define a GENOM_MODULE substitution in
PLISTs.

This feature is handled automatically and packages do not have to care about
this. All genom modules PLISTs can now be greatly simplified, but this is not
immediately required (duplicate entries are filtered correctly).

The print-PLIST target is aware of this and should generate the right content
for genom modules (i.e, just the extra files a module may install).